requirements.txt
# Requirements for Multi-Agent Visualization System

# Core LLM & Agent Framework
langchain==0.1.14
langchain-openai==0.1.1
langgraph==0.0.32
langsmith==0.1.31

# FastAPI & Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Data & Visualization
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Code Execution & Security
RestrictedPython==6.1
docker==7.0.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
